blink/bindings: Record whether the disk is usable for ParkableStrings.
ParkableStrings can be written out to disk, but this is not always the
case, as writing is disabled in several cases:
- When the renderer's associated profile is an incognito one
- When the browser failed to open the file
- When writing failed at some point

Capture the frequency of these events in a histogram. This requires
making the feature accessible to blink, so it moves from content to
blink.

Bug: 1029320
Change-Id: I671db1e92fcabc24c383551621ef93febd197cce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2249858
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779739}