[b4p/Settings] Add dialog for removing duplicated passwords
For a b4p user, in case some password has copies attached both to the
device and the Google Account, attempting to delete it should show a
dialog that allows selecting which copies to remove. This CL implements
this dialog but does not yet wire it in. Some details in the UI are left
for future CLs, such as the user avatar and displaying the website
associated with the password to be deleted. Note RemovePasswordBehavior
is not used here, for reasons that are better explained in
crrev.com/c/2246157.

Screenshot of implementation:
https://screenshot.googleplex.com/hZnTiuy80tL

Screenshot of mock:
https://screenshot.googleplex.com/pGV0iMTQBxw

Bug: 1049141
Change-Id: I50ca0e5bfc99dbc76c12c345c00f82788bbc377c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2248180
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779743}