[b4p/Settings] Display removal dialog
This CL starts displaying the dialog from crrev.com/c/2248180. Moreover,
RemovePasswordBehavior is deprecated. The more elaborate logic of
removing different copies of a password can live directly in the dialog,
while PasswordsListHandler can use simple logic for the case when a
single copy must be removed. In particular, it felt strange that
PasswordListItem has this Polymer behavior but never used any of its
functionalities (it was PasswordsListHandler that used it instead).

We also use the occasion to fix an aspect of crrev.com/c/2235850. If
the entry is moved, there's no element to reset the focus to, so the
focusWithoutInk() call is removed. The same is done for removal dialog.

Bug: 1049141
Change-Id: I06fe35b8bd14ee27873b6a56849e755c02cbc1bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2246157
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779757}