Show last day notification on user sign in
As per product requirements of minimum chrome version policy, it is
required to show notifications (for update required in case of network
limitations and to return the device in case of end-of-life) on login
if it is the last day to reach the deadline. Like other notifications
of this policy, this notification on sign in is also only applicable
to managed users to maintain consistency.

This behaviour helps to curb the problem of notification being missed
if policy is received when un-managed user is signed in and later
a managed user signs in. Only the last day notification is shown on
managed user sign in to avoid too much spamming.


Bug: 1048607
Change-Id: I01a8d52bbc4a003d5e1323140885881887211176
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2248567
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Commit-Queue: Saurabh Nijhara <snijhara@google.com>
Cr-Commit-Position: refs/heads/master@{#779770}