[Passwords][Leak Check] Scale spinner with font size
With larger (or smaller font) the absolute pixel sizes for the spinner
change the relative center of the animation which leads to an uneven
rotation.
To prevent this, the spinner now has a size relative to the font size
which is identical for the current size on normal scaling but scales
with larger/smaller sizes to keep the animation intact.

Since a large/small spinner looks wrong for a static status icon, the
icons also scale with the font size to match the spinner (which is not
a problem for the vector graphics).

Bug: 1096397
Change-Id: Ice9300cafdd2ef4877f28dd08c0a5b432588ddfb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2250203
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779771}