[b4p/metrics] Record number of unsynced passwords in the account store
This CL adds a histogram to record the number of account passwords that
are unsynced by the time the user signs out. We also use the occasion
to update the term "passwords account storage" to "account-scoped
password storage" in the description of another b4p histogram, and also
replace @chromium with @google in the owners emails.

Bug: 1063852
Change-Id: Ide376f40664112ada7ee62a404e019d53a2833bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2251826
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779797}