Make sure collapsed tabs are not visible.
Fixes a bug where a clipped favicon was sometimes visible in a collapsed
tab even when the tab is collapsed.

Tabs which are in collapsed groups *and* which are zero-width (so not
counting tabs which are animating to or from the collapsed state) now
cause TabStrip::ShouldTabBeVisible() to return false.

Bug: 1094051
Change-Id: Ie696a6e8442f6a4a761b7dee279da2e1a96ea1e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2251738
Commit-Queue: Charlene Yan <cyan@chromium.org>
Reviewed-by: Charlene Yan <cyan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779843}