WebUI Tab Strip: Optimize group move events
With the introduction of TabGroupChange::kMoved, there were multiple
tab-group-moved events being sent, one for the group moving and one
for the tabs in the group moving. This CL removes one of these events.

This CL also selects a group's tabs when the WebUI tab strip moves
the group to mirror the native tab strip and to allow the
TabStripModelChange::kMoved events of the group's tabs to be
ignored correctly. Previously, tab-moved events were being sent for
each tab when a group was moved using the WebUI tab strip which
caused for multiple unnecessary animations.

Bug: 1082344
Change-Id: I663d46e2e9a9857cdbaba8c17d2ff526f2a2fff8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2246956
Commit-Queue: John Lee <johntlee@chromium.org>
Reviewed-by: dpapad <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779856}