Remove TouchAction.* histograms
This CL removes TouchAction.* related histograms. The reason of
removing those is that we can't get useful info out from it.

Here is the doc for compositor touch action:
https://docs.google.com/document/d/1bSv_qNsLLnpXImuH7sHa6D3R_0GdDYkTiNjxso8Rgxo/edit

In the doc, there are 3 phases and currently we have phase 1
completed. Here is a short summary for phase 1 & 2:
phase 1: for GSU, if it is not allowed by allow lsited touch action,
then let's put it in a deferred gesture event queue, and wait for
the touch action comptued by blink to process that queue.

phase 2: for GSU, if the allow listed touch action is PanX
or PanY, then accumulate the delta on the y / x axis, and then
wait for the blink computed touch action to tell whether the
accumulated delta should be dropped or allowed. This isn't
carried out because it will have user-facing behaviro change.

So in order to evaluate how impactful phase 2 would be, we
need the following info: When a GSU is put in deferred
queue (which is only 0.01% of all GSUs), how often does that
GSU is allowed / filtered by the blink computed touch action.

Our current histograms don't give us that. Also, basically on
the amount of GSUs that are put in the deferred queue (0.01%),
it is probably not worth to carry out phase 2.

Bug: 1089591, 1089592
Change-Id: Iaa2e1011ac3fa1d4a124a60424d97b329b39647f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2246934
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Mustaq Ahmed <mustaq@chromium.org>
Commit-Queue: Xida Chen <xidachen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779860}