Add UMA histograms for skipped frames
This CL adds UMA histograms for the 1st, 2nd, and 3rd-5th animation
loops. The histogram tracks the number of frames skipped.

We expect to see a fairly large quantity of small skips in the 1st loop.
This has the effect of leaving holes in CC's cache.

The 2nd loop would then have a larger number of frames skipped
because filling the holes in the cache will cause multi-frame decodes.

The 3rd-5th loops should begin to normalize to 0 frames skipped.

Bug: 1094716
Change-Id: I85d22e1223d65bde398451fba1a9aff36698c114
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2245665
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Commit-Queue: Chris Blume <cblume@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779861}