Remove Unused RasterModeHint
In Canvas2DLayerBridge::ShouldAccelerate, the hint in the parameter is
never used. Furthermore, in GetOrCreateResourceProvider, the hint is
used to decide whether the creation should happen through restore or
not. The logic should be to use restore if the canvas is accelerated,
or create a new one otherwise. With a previous change to
ShouldAccelerate that preserves the acceleration mode, if the canvas
is currently using gpu and if the hint given is kPreferCPU, it would
still create a GPU canvas, in which case we could try to restore an
old one first. Therefore, the usage of the hint parameter can be
replaced by ShouldAccelerate(), and the parameter itself is not useful
after all.

All unused usages of RasterModeHint are cleaned up accordingly.

Bug: 1090180
Change-Id: If93cae51a567f7110085fa4514d0e60c56b07541
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2244115
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: James Darpinian <jdarpinian@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Teresa Kang <teresakang@google.com>
Cr-Commit-Position: refs/heads/master@{#779909}