When Serializing the CopyOutputRequest use the threadpool for expensive copies.
Currently when Chrome requests a screenshot we send a CopyOutputRequest
from the browser. However while taking this screenshot we need to
transfer it over a mojo pipe. This serialization takes 100-200 ms an
example traces in the linked bug on the VizCompositor thread and then
another 100-200 ms to deserialize it on the CrBrowserMain. This can
block high priority tasks which delay input handling causing visible
jank to users.

We fix this by doing the expensive copies on the ThreadPool, most other
copy tasks only take a couple ms so this removes most of the pain
relatively easily.

Tests had to be updated to know to wait for
CopyOutputResult::SendResult to complete async.


Bug: internal b/152858656
Change-Id: I114d65894c80742309bc2a02fdd2a0a7ad56dae7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2201762
Commit-Queue: Stephen Nusko <nuskos@chromium.org>
Reviewed-by: Eric Seckler <eseckler@chromium.org>
Reviewed-by: vmpstr <vmpstr@chromium.org>
Reviewed-by: Mitsuru Oshima (slow:gardening) <oshima@chromium.org>
Reviewed-by: Chris Palmer <palmer@chromium.org>
Reviewed-by: Stephen Nusko <nuskos@chromium.org>
Auto-Submit: Stephen Nusko <nuskos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779930}