Minimum chrome version notifications for EOL and metered network
This CL introduces the handling of showing in-session notifications
when update is required as per the policy but the device is
connected to a metered network or has reached end-of-life(EOL).

In case of metered network -
The notification is to be shown when the policy is received and on the
last day of the deadline. Clicking on the notification button starts
update process over metered connection.
For end-of-life devices -
The notification is to be shown when the policy is received, one week
before deadline and on the last day of the deadline. Clicking on the
notification button opens the enterprise/management info page which
would contain details about the end-of-life of the device.

This detail on the settings page will be implemented in subsequent CL.
New browser and unit tests are also included to verify this behavior.

Bug: 1048607
Change-Id: I4c09a00a50ff1673db1b15e4f05a4a29e825f043
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2245612
Commit-Queue: Saurabh Nijhara <snijhara@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779964}