WorkerTaskProvider: Don't unregister from ProfileManager if null
The ordering between the destruction of each WorkerTaskProvider and
the ProfileManager/g_browser_process instance is not guaranteed.
This means that it's necessary to check if the ProfileManager still
exists before unregistering from it, which preclude us from using
ScopedObserver.

This is a common pattern for other ProfileManager observers
instances.

Bug: 1094235
Change-Id: I35369d8938c95dcc6ef5204ea15b5f4dfa4d000e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2248073
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779982}