Skip EventLatency metrics if event time stamp is after cc stages
Ideally, if an event is associated with a compositor frame, it means
that the event was created before the compositor frame is submitted to
viz for presentation. So, the time stamp of the event should be smaller
than the start time of viz stage in CompositorFrameReporter. However, it
seems that it's not the case all the time (see the associated bug). For
now, let's silently ignore such cases instead of a DCHECK. Hopefully,
our work on investigating discrepancies between the new EventLatency and
the old Event.Latency would reveal the cause of the issue. If not, we
should reconsider this case.

Bug: 1093698
Change-Id: I4ce943fe2c9ed69caba82b34b652fa14e027d549
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2252678
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Commit-Queue: Mohsen Izadi <mohsen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#779997}