X11 and Ozone: fix use_x11 && use_ozone build for gl
When use_x11 && use_ozone are set to true at the same
time, compiler throws an error about multiple definitions
of the gl_factory and gl_initializer.

Thus, to fix that, make X11 use ozone's definition of the
gl_factory and the gl_initializer and add a check for
usage of the OzonePlatform that will be used in the future
when use_x11 && use_ozone are both set.

Please note that it's impossible to build use_x11 && use_ozone without
some hacks in PlatformCursor code. The changes to that are on their
way to upstream.

Bug: 1085700
Change-Id: Icc4d54c0910627c90869a7c41e3e709021d77b4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2247639
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780020}