Fix a regression when pinning an app, its icon becomes chopped
This is a regression caused by my earlier CL:
https://crrev.com/c/2227615. That CL exposed a bug in the
ordering between adding a view to the view model, and as
a child to ShelfView. The view should be added the model
first before adding it as a child to avoid have a wrong
view count in the model when updating the visible views.

Also, this CL fixes the value of
visible_shelf_item_bounds_union_ which was being updated
too early while views are animating making it have the
wrong bounds.

BUG=1095601
TEST=Manually, added a new test.

Change-Id: I5bca5a6febe06bfad791e2321727975bd1530d32
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2252931
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Andrew Xu <andrewxu@chromium.org>
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780059}