Add CdmStorage mojo interface for CdmFactoryDaemon
This is the mojo interface and implementation to adapt CdmStorage
requests from the Chrome OS daemon to the CdmStorage mojo interface in
Chrome. This will be passed into the CreateCdm mojo call in the
CdmFactory to provide origin specific storage for the Chrome OS CDM.

The implementation of the media::CdmFactory and
media::ContentDecryptionModule interfaces that us this will live
in the GPU process, which is why this component has the '_gpu'
suffix on it.

BUG=b:153111783
TEST=Unit test passes

Change-Id: I20e175e22f462a96a3c09facf06df7233484dc6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2252684
Commit-Queue: J Kardatzke <jkardatzke@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780103}