Allow ADSF construction with a provided activation state
AsyncDocumentSubresourceFilter (ADSF) initialization is always performed
asynchronously. As the activation state is currently computed during
initialization, the activation state cannot currently be accessed until
that process is complete.

Soon, we will construct an ADSF during DidFinishNavigation for certain
mainframes with an aborted load. Subsequent logic (e.g. for child
frames) requires that the activation state be immediately available. We
thus add a constructor that synchronously provides the activation state
for this case. To ensure the stability of the activation state, this
constructor should only be used when the ruleset is known to be
available and intact, for example, when a frame has an activated opener.

Bug: 1055558
Change-Id: I9ff235a0da82eecd47fd9b16e3da0e8f14e40e9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2240904
Commit-Queue: Alex Turner <alexmt@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780127}