Overview: Fix DeskNameView context menu closing overview
OverviewController used to handle OnAttemptToReactivateWindow()
by treating the window "requested" for activation as if it's the
window that gained activation (while actually activation didn't
change). This lead to overview closing when the requested window
for activation is a context menu (which is not activateable in the
first place).

It seems that OverviewController doesn't need to handle this
event at all, so removing.

BUG=1056541

Change-Id: Ic7d395bf448ca547b9173d5b153e55b90e3e5210
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2249125
Reviewed-by: Mitsuru Oshima (slow:gardening) <oshima@chromium.org>
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#780141}