Clamp percentage-based scroll deltas
This CL clamps the scrolling deltas for percent-based scrolling.

It is important for 2 cases:

- Small scrollers can have too small deltas, making scroll be
exaustingly slow. A minimum of 16 pixels can be set, as it is a
commonly used value for default font size in many browsers and could
match in many cases the text content inside the subscroller.

- Large scrollers can have excessively large scroll deltas, so it will
be limited by 12.5% of the viewport, that is the root scroller's
expected scroll delta. For OOPIFs content, the viewport dimensions
will be the iframe's.

This is done as a part of the effort to port and enhance Edge-style
scrolling into Chromium.

Bug: 1039047, 1074066

Change-Id: Ia4e0849eef3820fa3dc32b71c20257e63451ed4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2138493
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Rahul Arakeri <arakeri@microsoft.com>
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Commit-Queue: João Victor Almeida de Aguiar <joalmei@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#780153}