Split AXContentTreeData and AXContentTreeUpdate to separate files
Currently, AXContentTreeData and AXContentTreeUpdate are defined in
the same source files than AXContentNodeData (ax_content_node_data.h),
which makes it a bit unconvenient now that we are in the process of
removing all these classes, in favour of those from //ui/accessibility.

While technically this is not a requirement for the process, it would
be convenient to have each definition on separate places so that we
can work in a more iterative way towards removing those three classes,
so this CL puts AXContentTreeData in ax_content_tree_data.{h|cc} and
AXContentTreeupdate in ax_content_tree_update.h, in preparation for
the removal of ax_content_node_data.h, which will be the first step.

Bug: 1094150
Change-Id: Ied779d527fdff7dddeac896084872f4c15b31683
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2250603
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Cr-Commit-Position: refs/heads/master@{#780243}